Made Dataset.sample support Image.sample signature
